requirements.txt
pinecone
transformers
pymupdf
langchain
sentence_transformers
python-dotenv
protobuf
groq
scikit-learn
# fastapi
# uvicorn
# pymupdf
# requests
# python-dotenv
# langchain
# pinecone-client
